requirements.txt
requests==2.25.1
python-dotenv==1.0.0
argparse==1.4.0